requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.6
djangorestframework==3.11.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
greenlet==0.4.16
hiredis==1.0.1
httpie==2.1.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
imutils==0.5.3
incremental==17.5.0
kombu==4.6.8
msgpack==0.6.2
numpy==1.18.4
opencv-contrib-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2020.1
redis==3.5.2
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
websocket-client==0.57.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
